more cleanups

+++ db/migrate/20140121035618_move_rating_to_users.rb
@@ -1,6 +1,6 @@
 class MoveRatingToUsers < ActiveRecord::Migration
   def change
-    drop_table :ratings
+    drop_table :ratings if table_exists?(:ratings)
 
     add_column :users, :rating, :integer, default: 0
     add_column :users, :robustness, :integer, default: 0
